% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duncan.R
\name{duncan}
\alias{duncan}
\title{Multiple comparison: Duncan test}
\usage{
duncan(y, trt, DFerror, SSerror, alpha = 0.05, group = TRUE, main = NULL)
}
\arguments{
\item{y}{Numeric or complex vector containing the response
variable.}

\item{trt}{Numeric or complex vector containing the
treatments.}

\item{DFerror}{Error degrees of freedom.}

\item{SSerror}{Error sum of squares.}

\item{alpha}{Significance level.}

\item{group}{TRUE or FALSE.}

\item{main}{Title.}
}
\value{
Returns the multiple comparison of means according
to the test of Duncan.
}
\description{
\code{duncan} Performs the test of Duncan for multiple
comparison of means.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
}
