% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est21Ad.r
\docType{data}
\name{est21Ad}
\alias{est21Ad}
\title{Stink bugs in corn: additional treatment.}
\format{
Numeric vector.
}
\usage{
data(est21Ad)
}
\description{
Additional treatment response variable (height
of corn plants) of the experiment on stink bugs.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
