% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.r
\docType{data}
\name{ex}
\alias{ex}
\title{Vines: Split-Plot in Randomized Blocks Design}
\format{
A data frame with 24 observations on the following 4
variables.
  \describe{
  \item{\code{trat}}{a factor with levels \code{A} \code{B}}
  \item{\code{dose}}{a numeric vector}
  \item{\code{rep}}{a numeric vector}
  \item{\code{resp}}{a numeric vector}
  }
}
\usage{
data(ex)
}
\description{
Experiment about vines (not published) where one
studied the effects of different fertilizers and harvest dates
on the pH of grapes.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
