% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1.r
\docType{data}
\name{ex1}
\alias{ex1}
\title{Yacon: CRD}
\format{
A data frame with 24 observations on the following 2
variables.
  \describe{
  \item{\code{trat}}{a numeric vector}
  \item{\code{ig}}{a numeric vector}
  }
}
\usage{
data(ex1)
}
\description{
Experiment aiming to evaluate the influence of
the yacon flour consumption on the glicemic index.
}
\references{
RIBEIRO, J. de A. Estudos Quimicos e bioquimicos
do Yacon (Samallanthus sonchifolius) in natura e Processado e
Influencia do seu Consumo sobre Niveis Glicemicos e Lipideos
Fecais de Ratos. 2008. 166p. Dissertation (Master in Food
Science) - Universidade Federal de Lavras, UFLA, Lavras,
2008.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
