% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex2.r
\docType{data}
\name{ex2}
\alias{ex2}
\title{Food bars: RBD}
\format{
A data frame with 350 observations on the following
3 variables.
  \describe{
    \item{\code{provador}}{a numeric vector}
    \item{\code{trat}}{a factor with levels \code{A}
    \code{B} \code{C} \code{D} \code{E}}
    \item{\code{aparencia}}{a numeric vector}
  }
}
\usage{
data(ex2)
}
\description{
Sensory evaluation of food bars where panelists
(blocks) evaluated their appearance.
}
\references{
PAIVA, A. P. de. Estudos Tecnologicos, Quimico,
Fisico-quimico e Sensorial de Barras Alimenticias Elaboradas
com Subprodutos e Residuos Agoindustriais. 2008. 131p.
Dissertation (Master in Food Science) - Universidade Federal
de Lavras, UFLA, Lavras, 2008.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
