% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex3.r
\docType{data}
\name{ex3}
\alias{ex3}
\title{Forage: LSD}
\format{
A data frame with 49 observations on the following
4 variables.
  \describe{
  \item{\code{trat}}{a factor with levels \code{A} \code{B}
  \code{C} \code{D} \code{E} \code{F} \code{G}}
  \item{\code{linha}}{a numeric vector}
  \item{\code{coluna}}{a numeric vector}
  \item{\code{resp}}{a numeric vector}
  }
}
\usage{
data(ex3)
}
\description{
Data from an experiment aiming to select forage
for minimizing the intake problem of feeding cattle in the
sub-region of Paiaguas.
}
\references{
COMASTRI FILHO, J. A. Avaliacao de especies de
forrageiras nativas e exoticas na sub-regiao dos paiaguas no
pantanal mato-grossense. Pesq. Agropec. Bras., Brasilia,
v.29, n.6, p. 971-978, jun. 1994.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
