% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex5.r
\docType{data}
\name{ex5}
\alias{ex5}
\title{Food bars: Double Factorial scheme in RBD}
\format{
A data frame with 160 observations on the following
4 variables.
  \describe{
    \item{\code{trat}}{a factor with levels \code{10g}
    \code{15g} \code{15t} \code{20t}}
    \item{\code{genero}}{a factor with levels \code{F}
    \code{M}}
    \item{\code{bloco}}{a numeric vector}
    \item{\code{sabor}}{a numeric vector}
    }
}
\usage{
data(ex5)
}
\description{
Data adapted from a sensorial experiment where
panelists of different genders evaluated the taste of food
bars.
}
\references{
MOREIRA, D. K. T. Extrudados Expandidos de
Arroz, Soja e Gergelim para Uso em Barras Alimenticias.
2010. 166p. Dissertation (Master in Food Science) -
Universidade Federal de Lavras, UFLA, Lavras, 2010.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
