% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex6.r
\docType{data}
\name{ex6}
\alias{ex6}
\title{Fictional data 1}
\format{
A data frame with 24 observations on the following
5 variables.
  \describe{
   \item{\code{fatorA}}{a numeric vector}
   \item{\code{fatorB}}{a numeric vector}
   \item{\code{fatorC}}{a numeric vector}
   \item{\code{rep}}{a numeric vector}
   \item{\code{resp}}{a numeric vector}
   }
}
\usage{
data(ex6)
}
\description{
Data simulated from a standard normal
distribution for an experiment in triple factorial scheme.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
