% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex7.r
\docType{data}
\name{ex7}
\alias{ex7}
\title{Height of corn plants 21 days after emergence.}
\format{
Data frame with 80 observations on the following 4
variables.
  \describe{
   \item{\code{periodo}}{a factor with levels \code{0-7DAE}
   \code{0-14DAE} \code{0-21DAE} \code{7-14DAE}
   \code{7-21DAE}}
   \item{\code{nivel}}{a numeric vector}
   \item{\code{bloco}}{a numeric vector}
   \item{\code{est21}}{a numeric vector}
   }
   @references RODRIGUES, R. B. Danos do
   percevejo-barriga-verde Dichelops melacanthus (Dallas,
   1851) (Hemiptera: Pentatomidae) na cultura do milho.
   2011. 105f. Dissertacao (Mestrado em Agronomia -
   Universidade Federal de Santa Maria, Santa Maria, 2011.
}
\usage{
data(ex7)
}
\description{
We evaluated the height of corn plants 21 days
after emergence under infestation of stink bugs (Dichelops)
at different times of coexistence (period) and infestation
levels (level). Additional treatment is period zero and
level zero.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
