% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex8.r
\docType{data}
\name{ex8}
\alias{ex8}
\title{Composting: double factorial scheme plus one additional
treatment in CRD.}
\format{
A data frame with 24 observations on the following
5 variables.
  \describe{
 \item{\code{inoculante}}{a factor with levels
 \code{esterco} \code{mamona}}
 \item{\code{biodiesel}}{a numeric vector}
 \item{\code{vaso}}{a numeric vector}
 \item{\code{fresca}}{a numeric vector}
 \item{\code{seca}}{a numeric vector}
   }
}
\usage{
data(ex8)
}
\description{
Experiment in greenhouses to observe the
performance of the obtained composting for fertilizing
sorghum.
}
\references{
REZENDE, F. A. de. Aproveitamento  da Casca de
Cafe e Borra da Purificacao de Gorduras e Oleos Residuarios
em Compostagem. 2010. 74p. Thesis (Doctorate in
Agronomy/Fitotecny) - Universidade Federal de Lavras, UFLA,
Lavras, 2010.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
