% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex9.r
\docType{data}
\name{ex9}
\alias{ex9}
\title{Vegetated: Split-plot in CRD}
\format{
A data frame with 48 observations on the following
4 variables.
  \describe{
   \item{\code{cobertura}}{a factor with levels \code{T1}
   \code{T2} \code{T3} \code{T4} \code{T5} \code{T6}}
   \item{\code{prof}}{a numeric vector}
   \item{\code{rep}}{a numeric vector}
   \item{\code{pH}}{a numeric vector}
   }
}
\usage{
data(ex9)
}
\description{
Subset of data from an experiment that studied
the effect on soil pH of cover crops subjected to trampling
by cattle predominantly under continuous grazing system,
analyzed at different depths.
}
\references{
GUERRA, A. R. Atributos de Solo sob Coberturas
Vegetais em Sistema Silvipastoril em Lavras - MG. 2010. 141p.
Dissertation (Master in Forest Engineering) - Universidade
Federal de Lavras, UFLA, Lavras, 2010.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
