% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exnl.r
\docType{data}
\name{exnl}
\alias{exnl}
\title{Example of fictitious data set}
\format{
A data frame with 30 observations of the following
3 variables.
  \describe{
  \item{\code{trat}}{a numeric vector}
  \item{\code{rep}}{a numeric vector}
  \item{\code{resp}}{a numeric vector}
  }
}
\usage{
data(exnl)
}
\description{
Example of fictitious data mass for non-linear
regression model fit
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
