% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/han.R
\name{han}
\alias{han}
\title{Test for homogeneity of variances of Han}
\usage{
han(resp, trat, block)
}
\arguments{
\item{resp}{Numeric or complex vector containing the response
variable.}

\item{trat}{Numeric or complex vector containing the
treatments.}

\item{block}{Numeric or complex vector containing the blocks.}
}
\value{
Returns the p-value of Han's test of homogeneity of
variances and its practical interpretation for 5\% of
significance.
}
\description{
\code{han} Performs the test for homogeneity of variances of
Han (1969).
}
\examples{
data(ex2)
attach(ex2)
rbd(trat, provador, aparencia, hvar = "han")
}
\references{
HAN, C. P. Testing the homogeneity of variances
in a two-way classification. \emph{Biometrics}, 25:153-158,
Mar. 1969.

RIBEIRO, R. \emph{Proposta e comparacao do desempenho de
testes para homogeneidade de variancia de modelos de
classicacao one-way e two-way}. Iniciacao Cientifica.
(Iniciacao Cientifica) - Universidade Federal de Alfenas.
2012.
}
\seealso{
\code{\link{anscombetukey}},
\code{\link{oneillmathews}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Marcos Costa de Paula
 @author Mateus Pimenta Siqueira Lima
}
