% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneilldbc.R
\name{oneilldbc}
\alias{oneilldbc}
\title{Test for homogeneity of variances of ONeill and Mathews (RBD)}
\usage{
oneilldbc(resp, trat, block)
}
\arguments{
\item{resp}{Numeric or complex vector containing the response
variable.}

\item{trat}{Numeric or complex vector containing treatments.}

\item{block}{Numeric or complex vector containing blocks.}
}
\value{
Returns the p-value of ONeill and Mathews' test of
homogeneity of variances and its practical interpretation
for significance level of 5\%.
}
\description{
\code{oneilldbc} Performs the test for homogeneity of
variances of ONeill and Mathews (2002).
}
\examples{
data(ex2)
attach(ex2)
rbd(trat, provador, aparencia, hvar = "oneillmathews")
}
\references{
O'NEILL, M. E.; MATHEWS, K. L. Levene tests of
homogeneity of variance for general block and treatment
designs. \emph{Biometrics}, 58:216-224, Mar. 2002.

RIBEIRO, R. \emph{Proposta e comparacao do desempenho de
testes para homogeneidade de variancia de modelos de
classificacao one-way e two-way}. Iniciacao Cientifica.
(Iniciacao Cientifica) - Universidade Federal de Alfenas.
2012.
}
\seealso{
\code{\link{anscombetukey}}, \code{\link{han}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Marcos Costa de Paula
 @author Mateus Pimenta Siqueira Lima
}
