% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotres.R
\name{plotres}
\alias{plotres}
\title{Residual plots}
\usage{
plotres(x)
}
\arguments{
\item{x}{Output from anova (performed in ExpDes).}
}
\description{
\code{plotres} Residual plots for a output model. Four sets
of plots are produced: (1) Histogram, (2) normal probability
plot for the residual, (3) Standardized Residuals versus
Fitted Values, and (4) box-plot (Standardized Residuals).
}
\examples{
data(ex1)
attach(ex1)
a<-crd(trat, ig)
plotres(a)
}
\references{
STEEL, R. G. D.; TORRIE, J. H. \emph{Principles
and procedures in Statistics: a biometrical approach}.
McGraw-Hill, New York, NY. 1980.
}
\seealso{
\code{\link{graphics}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @note The default produces four plots regarding the ANOVA
 assumptions.
}
