% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.nl.R
\name{reg.nl}
\alias{reg.nl}
\title{Non-linear Regression}
\usage{
reg.nl(resp, treat)
}
\arguments{
\item{resp}{Numeric or complex vector containing the
response variable.}

\item{treat}{Numeric or complex vector containing the
treatments.}
}
\value{
Returns coefficients, significance and ANOVA of the
fitted regression models.
}
\description{
\code{reg.nl} Adjusts non-linear regression models in Anova
(Models: Power, Exponential, Logistic, Gompertz).
}
\examples{
data(exnl)
attach(exnl)
x<-crd(trat, resp, quali = FALSE, nl = TRUE)
graphics(x, degree = "log")
}
\references{
DRAPER, N.R.; SMITH, H. \emph{Apllied regression
analysis}. 3ed. New York : John Wiley, 1998. 706p.
}
\seealso{
\code{\link{graphics}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Luiz Alberto Beijo
}
