% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.poly.R
\name{reg.poly}
\alias{reg.poly}
\title{Polinomial Regression}
\usage{
reg.poly(resp, treat, DFerror, SSerror, DFtreat, SStreat)
}
\arguments{
\item{resp}{Numeric or complex vector containing the
response variable.}

\item{treat}{Numeric or complex vector containing the
treatments.}

\item{DFerror}{Error degrees of freedom.}

\item{SSerror}{Error sum of squares.}

\item{DFtreat}{Treatments' dregrees of freedom.}

\item{SStreat}{Treatments' sum of squares.}
}
\value{
Returns coefficients, significance and ANOVA of the
fitted regression models.
}
\description{
\code{reg.poly} Fits sequential regression models until the
third power.
}
\references{
GOMES, F. P. Curso de Estatistica Experimental.
10a ed. Piracicaba: ESALQ/USP. 1982. 430.
}
\seealso{
\code{\link{graphics}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
}
