% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respAd.r
\docType{data}
\name{respAd}
\alias{respAd}
\title{Fictional data: additional treatment}
\format{
Numeric vector.
}
\usage{
data(respAd)
}
\description{
Response variable form the additional treatment.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
