% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scottknott.R
\name{scottknott}
\alias{scottknott}
\title{Multiple comparison: Scott-Knott test}
\usage{
scottknott(y, trt, DFerror, SSerror, alpha = 0.05, group = TRUE, main = NULL)
}
\arguments{
\item{y}{Numeric or complex vector containing the response
variable.}

\item{trt}{Numeric or complex vector containing the
treatments.}

\item{DFerror}{Error degrees of freedom.}

\item{SSerror}{Error sum of squares.}

\item{alpha}{Significance level.}

\item{group}{TRUE or FALSE.}

\item{main}{Title.}
}
\value{
Returns the multiple comparison of means according
to the test of Scott-Knott.
}
\description{
\code{scottknott} Performs the test of Scott-Knott, for
multiple comparison of means.
}
\examples{
data(ex1)
attach(ex1)
crd(trat, ig, quali = TRUE, mcomp = "sk", sigT = 0.05)
}
\references{
RAMALHO, M. A. P.; FERREIRA, D. F.; OLIVEIRA,
A. C. de. Experimentacao em Genetica e Melhoramento de
Plantas. 2a ed. Lavras: UFLA. 2005. 300p.
}
\seealso{
\code{\link{snk}}, \code{\link{duncan}},
\code{\link{lsd}}, \code{\link{lsdb}}, \code{\link{ccboot}},
\code{\link{tukey}}, \code{\link{ccF}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
(Adapted from Laercio Junio da Silva - GPL(>=2))
}
