% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secaAd.r
\docType{data}
\name{secaAd}
\alias{secaAd}
\title{Composting: additional treatment}
\format{
Numeric vector.
}
\usage{
data(secaAd)
}
\description{
Response variable (dry biomass) of the additional
treatment of the experiment about composting.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
