% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.R
\name{tukey}
\alias{tukey}
\title{Multiple comparison: Tukey's test}
\usage{
tukey(y, trt, DFerror, SSerror, alpha = 0.05, group = TRUE, main = NULL)
}
\arguments{
\item{y}{Numeric or complex vector containing the response
variable.}

\item{trt}{Numeric or complex vector containing the
treatments.}

\item{DFerror}{Error degrees of freedom.}

\item{SSerror}{Error sum of squares.}

\item{alpha}{Significance level.}

\item{group}{TRUE or FALSE.}

\item{main}{Title.}
}
\value{
y  Numeric
trt  factor
DFerror  Numeric
MSerror  Numeric
alpha  Numeric
group Logic
main  Text
}
\description{
\code{tukey} Performs the test of Tukey, for multiple
comparison of means.
}
\details{
It is necessary first makes a analysis of variance.
}
\examples{
data(ex1)
attach(ex1)
crd(trat, ig, quali = TRUE, mcomp = "tukey", sigT = 0.05)
}
\references{
Principles and procedures of statistics a
biometrical approach Steel and Torry and Dickey. Third
Edition 1997
}
\seealso{
\code{\link{scottknott}}, \code{\link{duncan}},
\code{\link{lsd}}, \code{\link{lsdb}}, \code{\link{ccboot}},
\code{\link{snk}}, \code{\link{ccF}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
(Adapted from Felipe de Mendiburu - GPL)
}
